major design changes, UI interface, moved searchbar to sidenav and moved state to app
