+++ public/css/styles.css
@@ -10,3 +10,37 @@ html, body {
 .ui.input {
   width: 100%;
 }
+
+.card.result:hover {
+  opacity: 0.8;
+}
+
+ #appContainer {
+  height: 2000px !important;
+  margin-top: 50px;
+ }
+
+ #navbar {
+  position: fixed;
+  top: 0;
+  left: 0;
+  /*border-bottom: 2px solid orange;*/
+  width: 100%;
+  height: 43px;
+  z-index: 1;
+ }
+
+ .nav-item {
+  border: 1px solid blue;
+  display: inline-block;
+ }
+
+.fluid.ui.button.header {
+  margin: 0px;
+  border-radius: 0;
+  height: 36px;
+}
+
+.header.item::before {
+  display: none;
+}
