+++ public/js/components/reciperesult.js
@@ -1,0 +1,68 @@
+'use strict'
+const React = require('react');
+
+
+const RecipeResult = React.createClass({
+
+  componentDidMount : function() {
+    $(".rating").rating();
+    console.log('reciperesults mounted');
+  },
+  handleClick : function() {
+    $('.ui.modal.'+this.props.index)
+    .modal('show')
+  },
+  handleSelect : function() {
+    
+  },
+  render : function() {
+    return (
+
+        <div className="card result" onClick={this.handleClick}>
+          <div className="image">
+            <img src={this.props.details.largeImage} className="recipeImg" />
+          </div>
+          <div className="extra">
+            Rating:
+            <div className="ui star rating" data-rating={this.props.details.rating}></div><br/>
+            {this.props.details.recipeName}
+          </div>
+
+          <div className={"ui modal "+this.props.index}>
+            <i className="close icon"></i>
+            <div className="header">
+              {this.props.details.recipeName}
+            </div>
+            <div className="image content">
+              <div className="ui medium image">
+                <img src={this.props.details.largeImage} />
+              </div>
+              <div className="description">
+                <div className="ui header">Ingredients</div>
+                <ul>
+                  {
+                    this.props.details.ingredients.map(function(el) {
+                      return (<li>{el}</li>)
+                    })
+                  }
+                </ul>
+                <p>Total time: {this.props.details.totalTime}</p>
+              </div>
+            </div>
+            <div className="actions">
+              <div className="ui black deny button">Nope
+                
+              </div>
+              <div className="ui positive right labeled icon button" onClick={this.handleSelect}>
+                Cook it
+                <i className="checkmark icon"></i>
+              </div>
+            </div>
+          </div>
+
+        </div>
+    )
+  }
+});
+
+module.exports = RecipeResult;
