+++ routes/users.js
@@ -19,11 +19,13 @@ users.post('/yummly', (req, res) => {
   let allowed = req.body.qs.split('+');
   allowed = allowed.join('&allowedIngredient[]=');
 
-  request(`http:/\/api.yummly.com/v1/api/recipes?_app_id=${process.env.YUMMLY_APP_ID}&_app_key=${process.env.YUMMLY_APIKEY}&allowedIngredient[]=${allowed}&maxResult=16`, function (error, response, body) {
+  request(`http:/\/api.yummly.com/v1/api/recipes?_app_id=${process.env.YUMMLY_APP_ID}&_app_key=${process.env.YUMMLY_APIKEY}&allowedIngredient[]=${allowed}&maxResult=1`, function (error, response, body) {
     if (!error && response.statusCode == 200) {
       res.send(body);
     }
   });
+
+  // res.send('Out of order');
 })
 
 users.get('/yummly/:id', (req, res) => {
@@ -33,6 +35,7 @@ users.get('/yummly/:id', (req, res) => {
       res.send(body);
     }
   });
+  // res.send('Out of order');
 })
